Rename PilotAgent directory to Pilot
Also, modify import paths in unit tests
